infrastructure/requirements.txt
aws-cdk-lib==2.31.1
constructs>=10.0.0,<11.0.0
cdk_lambda_layer_builder @ git+https://github.com/aws-samples/aws-cdk-lambda-layer-builder.git#egg=cdk_lambda_layer_builder